releng_docs: updating dependencies

+++ src/releng_docs/requirements.nix
@@ -1,4 +1,4 @@
-# generated using pypi2nix tool (version: 1.8.0)
+# generated using pypi2nix tool (version: 1.8.1)
 # See more at: https://github.com/garbas/pypi2nix
 #
 # COMMAND:
@@ -23,7 +23,10 @@ let
       self: super: {
         bootstrapped-pip = super.bootstrapped-pip.overrideDerivation (old: {
           patchPhase = old.patchPhase + ''
-            sed -i               -e "s|paths_to_remove.remove(auto_confirm)|#paths_to_remove.remove(auto_confirm)|"                -e "s|self.uninstalled = paths_to_remove|#self.uninstalled = paths_to_remove|"                  $out/${pkgs.python35.sitePackages}/pip/req/req_install.py
+            sed -i \
+              -e "s|paths_to_remove.remove(auto_confirm)|#paths_to_remove.remove(auto_confirm)|"  \
+              -e "s|self.uninstalled = paths_to_remove|#self.uninstalled = paths_to_remove|"  \
+                $out/${pkgs.python35.sitePackages}/pip/req/req_install.py
           '';
         });
       };
@@ -40,11 +43,13 @@ let
         buildInputs = [ makeWrapper ] ++ (builtins.attrValues pkgs);
         buildCommand = ''
           mkdir -p $out/bin
-          ln -s ${pythonPackages.python.interpreter}               $out/bin/${pythonPackages.python.executable}
-          for dep in ${builtins.concatStringsSep " "               (builtins.attrValues pkgs)}; do
+          ln -s ${pythonPackages.python.interpreter} \
+              $out/bin/${pythonPackages.python.executable}
+          for dep in ${builtins.concatStringsSep " "
+              (builtins.attrValues pkgs)}; do
             if [ -d "$dep/bin" ]; then
               for prog in "$dep/bin/"*; do
-                if [ -f $prog ]; then
+                if [ -x "$prog" ] && [ -f "$prog" ]; then
                   ln -s $prog $out/bin/`basename $prog`
                 fi
               done
@@ -55,6 +60,8 @@ let
           done
           pushd $out/bin
           ln -s ${pythonPackages.python.executable} python
+          ln -s ${pythonPackages.python.executable} \
+              python3
           popd
         '';
         passthru.interpreter = pythonPackages.python;
@@ -65,7 +72,9 @@ let
       mkDerivation = pythonPackages.buildPythonPackage;
       packages = pkgs;
       overrideDerivation = drv: f:
-        pythonPackages.buildPythonPackage (drv.drvAttrs // f drv.drvAttrs);
+        pythonPackages.buildPythonPackage (
+          drv.drvAttrs // f drv.drvAttrs // { meta = drv.meta; }
+        );
       withPackages = pkgs'':
         withPackages (pkgs // pkgs'');
     };
@@ -73,10 +82,9 @@ let
   python = withPackages {};
 
   generated = self: {
-
     "Babel" = python.mkDerivation {
-      name = "Babel-2.4.0";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/92/22/643f3b75f75e0220c5ef9f5b72b619ccffe9266170143a4821d4885198de/Babel-2.4.0.tar.gz"; sha256 = "8c98f5e5f8f5f088571f2c6bd88d530e331cbbcb95a7311a0db69d3dca7ec563"; };
+      name = "Babel-2.5.3";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/0e/d5/9b1d6a79c975d0e9a32bd337a1465518c2519b14b214682ca9892752417e/Babel-2.5.3.tar.gz"; sha256 = "8ce4cb6fdd4393edd323227cba3a077bceb2a6ce5201c902c65e730046f41f14"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [
@@ -89,11 +97,9 @@ let
       };
     };
 
-
-
     "Jinja2" = python.mkDerivation {
-      name = "Jinja2-2.9.6";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/90/61/f820ff0076a2599dd39406dcb858ecb239438c02ce706c8e91131ab9c7f1/Jinja2-2.9.6.tar.gz"; sha256 = "ddaa01a212cd6d641401cb01b605f4a4d9f37bfc93043d7f760ec70fb99ff9ff"; };
+      name = "Jinja2-2.10";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/56/e6/332789f295cf22308386cf5bbd1f4e00ed11484299c5d7383378cf48ba47/Jinja2-2.10.tar.gz"; sha256 = "f84be1bb0040caca4cea721fcbbbbd61f9be9464ca236387158b0feea01914a4"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [
@@ -106,8 +112,6 @@ let
         description = "A small but fast and easy to use stand-alone template engine written in pure python.";
       };
     };
-
-
 
     "MarkupSafe" = python.mkDerivation {
       name = "MarkupSafe-1.0";
@@ -122,24 +126,18 @@ let
       };
     };
 
-
-
     "Pillow" = python.mkDerivation {
-      name = "Pillow-4.2.1";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/55/aa/f7f983fb72710a9daa4b3374b7c160091d3f94f5c09221f9336ade9027f3/Pillow-4.2.1.tar.gz"; sha256 = "c724f65870e545316f9e82e4c6d608ab5aa9dd82d5185e5b2e72119378740073"; };
+      name = "Pillow-5.0.0";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/0f/57/25be1a4c2d487942c3ed360f6eee7f41c5b9196a09ca71c54d1a33c968d9/Pillow-5.0.0.tar.gz"; sha256 = "12f29d6c23424f704c66b5b68c02fe0b571504459605cfe36ab8158359b0e1bb"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
-      propagatedBuildInputs = [
-      self."olefile"
-    ];
+      propagatedBuildInputs = [ ];
       meta = with pkgs.stdenv.lib; {
         homepage = "https://python-pillow.org";
-        license = "Standard PIL License";
+        license = "License :: Other/Proprietary License";
         description = "Python Imaging Library (Fork)";
       };
     };
-
-
 
     "Pygments" = python.mkDerivation {
       name = "Pygments-2.2.0";
@@ -154,11 +152,9 @@ let
       };
     };
 
-
-
     "Sphinx" = python.mkDerivation {
-      name = "Sphinx-1.6.3";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/10/91/ceb2e0d763e0c626f7afd7e3272a5bb76dd06eed1f0b908270ea31984062/Sphinx-1.6.3.tar.gz"; sha256 = "af8bdb8c714552b77d01d4536e3d6d2879d6cb9d25423d29163d5788e27046e6"; };
+      name = "Sphinx-1.7.1";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/b1/21/b97c0936171669a90e0e7036a112c55fc415699e8707e675b987949221e1/Sphinx-1.7.1.tar.gz"; sha256 = "da987de5fcca21a4acc7f67a86a363039e67ac3e8827161e61b91deb131c0ee8"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [
@@ -168,6 +164,7 @@ let
       self."alabaster"
       self."docutils"
       self."imagesize"
+      self."packaging"
       self."requests"
       self."six"
       self."snowballstemmer"
@@ -179,8 +176,6 @@ let
         description = "Python documentation generator";
       };
     };
-
-
 
     "actdiag" = python.mkDerivation {
       name = "actdiag-0.5.4";
@@ -198,8 +193,6 @@ let
       };
     };
 
-
-
     "alabaster" = python.mkDerivation {
       name = "alabaster-0.7.10";
       src = pkgs.fetchurl { url = "https://pypi.python.org/packages/d0/a5/e3a9ad3ee86aceeff71908ae562580643b955ea1b1d4f08ed6f7e8396bd7/alabaster-0.7.10.tar.gz"; sha256 = "37cdcb9e9954ed60912ebc1ca12a9d12178c26637abdf124e3cde2341c257fe0"; };
@@ -212,8 +205,6 @@ let
         description = "A configurable sidebar-enabled Sphinx theme";
       };
     };
-
-
 
     "blockdiag" = python.mkDerivation {
       name = "blockdiag-1.5.3";
@@ -233,22 +224,18 @@ let
       };
     };
 
-
-
     "certifi" = python.mkDerivation {
-      name = "certifi-2017.7.27.1";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/20/d0/3f7a84b0c5b89e94abbd073a5f00c7176089f526edb056686751d5064cbd/certifi-2017.7.27.1.tar.gz"; sha256 = "40523d2efb60523e113b44602298f0960e900388cf3bb6043f645cf57ea9e3f5"; };
+      name = "certifi-2018.1.18";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/15/d4/2f888fc463d516ff7bf2379a4e9a552fef7f22a94147655d9b1097108248/certifi-2018.1.18.tar.gz"; sha256 = "edbc3f203427eef571f79a7692bb160a2b0f7ccaa31953e99bd17e307cf63f7d"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [ ];
       meta = with pkgs.stdenv.lib; {
         homepage = "http://certifi.io/";
-        license = "MPL-2.0";
+        license = licenses.mpl20;
         description = "Python package for providing Mozilla's CA Bundle.";
       };
     };
-
-
 
     "chardet" = python.mkDerivation {
       name = "chardet-3.0.4";
@@ -263,11 +250,9 @@ let
       };
     };
 
-
-
     "docutils" = python.mkDerivation {
-      name = "docutils-0.13.1";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/05/25/7b5484aca5d46915493f1fd4ecb63c38c333bd32aa9ad6e19da8d08895ae/docutils-0.13.1.tar.gz"; sha256 = "718c0f5fb677be0f34b781e04241c4067cbd9327b66bdd8e763201130f5175be"; };
+      name = "docutils-0.14";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/84/f4/5771e41fdf52aabebbadecc9381d11dea0fa34e4759b4071244fa094804c/docutils-0.14.tar.gz"; sha256 = "51e64ef2ebfb29cae1faa133b3710143496eca21c530f3f71424d77687764274"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [ ];
@@ -277,8 +262,6 @@ let
         description = "Docutils -- Python Documentation Utilities";
       };
     };
-
-
 
     "funcparserlib" = python.mkDerivation {
       name = "funcparserlib-0.3.6";
@@ -293,11 +276,9 @@ let
       };
     };
 
-
-
     "idna" = python.mkDerivation {
-      name = "idna-2.5";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/d8/82/28a51052215014efc07feac7330ed758702fc0581347098a81699b5281cb/idna-2.5.tar.gz"; sha256 = "3cb5ce08046c4e3a560fc02f138d0ac63e00f8ce5901a56b32ec8b7994082aab"; };
+      name = "idna-2.6";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/f4/bd/0467d62790828c23c47fc1dfa1b1f052b24efdf5290f071c7a91d0d82fd3/idna-2.6.tar.gz"; sha256 = "2c6a5de3089009e3da7c5dde64a141dbc8551d5b7f6cf4ed7c2568d0cc520a8f"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [ ];
@@ -308,11 +289,9 @@ let
       };
     };
 
-
-
     "imagesize" = python.mkDerivation {
-      name = "imagesize-0.7.1";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/53/72/6c6f1e787d9cab2cc733cf042f125abec07209a58308831c9f292504e826/imagesize-0.7.1.tar.gz"; sha256 = "0ab2c62b87987e3252f89d30b7cedbec12a01af9274af9ffa48108f2c13c6062"; };
+      name = "imagesize-1.0.0";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/c6/3f/1db2da33804e8d7ef3a868b27b7bdc1aae6a4f693f0162d2aeeaf503864f/imagesize-1.0.0.tar.gz"; sha256 = "5b326e4678b6925158ccc66a9fa3122b6106d7c876ee32d7de6ce59385b96315"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [ ];
@@ -322,8 +301,6 @@ let
         description = "Getting image size from png/jpeg/jpeg2000/gif file";
       };
     };
-
-
 
     "livereload" = python.mkDerivation {
       name = "livereload-2.5.1";
@@ -341,8 +318,6 @@ let
       };
     };
 
-
-
     "nwdiag" = python.mkDerivation {
       name = "nwdiag-1.0.4";
       src = pkgs.fetchurl { url = "https://pypi.python.org/packages/ca/5e/8b434f6655869c31b0b6d47f3972a469dfe14cb7f5b3b3b7c7413fa08f1c/nwdiag-1.0.4.tar.gz"; sha256 = "002565875559789a2dfc5f578c07abdf44269c3f7cdf78d4809bdc4bdc2213fa"; };
@@ -359,26 +334,38 @@ let
       };
     };
 
+    "packaging" = python.mkDerivation {
+      name = "packaging-17.0";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/6d/72/20bcaab4f7a6bfd108c4ca1dc486906817efa9596e3b15b4c0603da981ba/packaging-17.0.tar.gz"; sha256 = "e9f654a6854321ac39d2e6745b820773ba9efa394e71dea1b387cc717d439f93"; };
+      doCheck = commonDoCheck;
+      buildInputs = commonBuildInputs;
+      propagatedBuildInputs = [
+      self."pyparsing"
+      self."six"
+    ];
+      meta = with pkgs.stdenv.lib; {
+        homepage = "https://github.com/pypa/packaging";
+        license = licenses.bsdOriginal;
+        description = "Core utilities for Python packages";
+      };
+    };
 
-
-    "olefile" = python.mkDerivation {
-      name = "olefile-0.44";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/35/17/c15d41d5a8f8b98cc3df25eb00c5cee76193114c78e5674df6ef4ac92647/olefile-0.44.zip"; sha256 = "61f2ca0cd0aa77279eb943c07f607438edf374096b66332fae1ee64a6f0f73ad"; };
+    "pyparsing" = python.mkDerivation {
+      name = "pyparsing-2.2.0";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/3c/ec/a94f8cf7274ea60b5413df054f82a8980523efd712ec55a59e7c3357cf7c/pyparsing-2.2.0.tar.gz"; sha256 = "0832bcf47acd283788593e7a0f542407bd9550a55a8a8435214a1960e04bcb04"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [ ];
       meta = with pkgs.stdenv.lib; {
-        homepage = "https://www.decalage.info/python/olefileio";
-        license = licenses.bsdOriginal;
-        description = "Python package to parse, read and write Microsoft OLE2 files (Structured Storage or Compound Document, Microsoft Office) - Improved version of the OleFileIO module from PIL, the Python Image Library.";
+        homepage = "http://pyparsing.wikispaces.com/";
+        license = licenses.mit;
+        description = "Python parsing module";
       };
     };
 
-
-
     "pytz" = python.mkDerivation {
-      name = "pytz-2017.2";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/a4/09/c47e57fc9c7062b4e83b075d418800d322caa87ec0ac21e6308bd3a2d519/pytz-2017.2.zip"; sha256 = "f5c056e8f62d45ba8215e5cb8f50dfccb198b4b9fbea8500674f3443e4689589"; };
+      name = "pytz-2018.3";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/1b/50/4cdc62fc0753595fc16c8f722a89740f487c6e5670c644eb8983946777be/pytz-2018.3.tar.gz"; sha256 = "410bcd1d6409026fbaa65d9ed33bf6dd8b1e94a499e32168acfc7b332e4095c0"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [ ];
@@ -389,11 +376,9 @@ let
       };
     };
 
-
-
     "requests" = python.mkDerivation {
-      name = "requests-2.18.2";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/07/2e/81fdfdfac91cf3cb2518fb149ac67caf0e081b485eab68e9aee63396f7e8/requests-2.18.2.tar.gz"; sha256 = "5b26fcc5e72757a867e4d562333f841eddcef93548908a1bb1a9207260618da9"; };
+      name = "requests-2.18.4";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/b0/e1/eab4fc3752e3d240468a8c0b284607899d2fbfb236a56b7377a329aa8d09/requests-2.18.4.tar.gz"; sha256 = "9c443e7324ba5b85070c4a818ade28bfabedf16ea10206da1132edaa6dda237e"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [
@@ -408,8 +393,6 @@ let
         description = "Python HTTP for Humans.";
       };
     };
-
-
 
     "seqdiag" = python.mkDerivation {
       name = "seqdiag-0.9.5";
@@ -427,11 +410,9 @@ let
       };
     };
 
-
-
     "six" = python.mkDerivation {
-      name = "six-1.10.0";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/b3/b2/238e2590826bfdd113244a40d9d3eb26918bd798fc187e2360a8367068db/six-1.10.0.tar.gz"; sha256 = "105f8d68616f8248e24bf0e9372ef04d3cc10104f1980f54d57b2ce73a5ad56a"; };
+      name = "six-1.11.0";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/16/d8/bc6316cf98419719bd59c91742194c111b6f2e85abac88e496adefaf7afe/six-1.11.0.tar.gz"; sha256 = "70e8a77beed4562e7f14fe23a786b54f6296e34344c23bc42f07b15018ff98e9"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [ ];
@@ -441,8 +422,6 @@ let
         description = "Python 2 and 3 compatibility utilities";
       };
     };
-
-
 
     "snowballstemmer" = python.mkDerivation {
       name = "snowballstemmer-1.2.1";
@@ -456,8 +435,6 @@ let
         description = "This package provides 16 stemmer algorithms (15 + Poerter English stemmer) generated from Snowball algorithms.";
       };
     };
-
-
 
     "sphinxcontrib-actdiag" = python.mkDerivation {
       name = "sphinxcontrib-actdiag-0.8.5";
@@ -476,8 +453,6 @@ let
       };
     };
 
-
-
     "sphinxcontrib-blockdiag" = python.mkDerivation {
       name = "sphinxcontrib-blockdiag-1.5.5";
       src = pkgs.fetchurl { url = "https://pypi.python.org/packages/04/50/7a43117a5a8a16acaceabc5ad69092fa1dacb11ef83c84fdf234e5a3502f/sphinxcontrib-blockdiag-1.5.5.tar.gz"; sha256 = "7cdff966d8f372b9536374954314a6cf4280e0e48bc2321a4f25cc7f2114f8f0"; };
@@ -493,8 +468,6 @@ let
         description = "Sphinx \"blockdiag\" extension";
       };
     };
-
-
 
     "sphinxcontrib-nwdiag" = python.mkDerivation {
       name = "sphinxcontrib-nwdiag-0.9.5";
@@ -513,8 +486,6 @@ let
       };
     };
 
-
-
     "sphinxcontrib-seqdiag" = python.mkDerivation {
       name = "sphinxcontrib-seqdiag-0.8.5";
       src = pkgs.fetchurl { url = "https://pypi.python.org/packages/10/59/5f746c6fe8a83ed7451b59e7787080adad8850a8a04d610713466fca3bca/sphinxcontrib-seqdiag-0.8.5.tar.gz"; sha256 = "83c3fdac7e083c5b217f65359c03b75af753209028db6b261b196aff19e7003f"; };
@@ -532,8 +503,6 @@ let
       };
     };
 
-
-
     "sphinxcontrib-websupport" = python.mkDerivation {
       name = "sphinxcontrib-websupport-1.0.1";
       src = pkgs.fetchurl { url = "https://pypi.python.org/packages/c5/6b/f0630436b931ad4f8331a9399ca18a7d447f0fcc0c7178fb56b1aee68d01/sphinxcontrib-websupport-1.0.1.tar.gz"; sha256 = "7a85961326aa3a400cd4ad3c816d70ed6f7c740acd7ce5d78cd0a67825072eb9"; };
@@ -547,11 +516,9 @@ let
       };
     };
 
-
-
     "tornado" = python.mkDerivation {
-      name = "tornado-4.5.1";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/df/42/a180ee540e12e2ec1007ac82a42b09dd92e5461e09c98bf465e98646d187/tornado-4.5.1.tar.gz"; sha256 = "db0904a28253cfe53e7dedc765c71596f3c53bb8a866ae50123320ec1a7b73fd"; };
+      name = "tornado-4.5.3";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/e3/7b/e29ab3d51c8df66922fea216e2bddfcb6430fb29620e5165b16a216e0d3c/tornado-4.5.3.tar.gz"; sha256 = "6d14e47eab0e15799cf3cdcc86b0b98279da68522caace2bd7ce644287685f0a"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [ ];
@@ -561,8 +528,6 @@ let
         description = "Tornado is a Python web framework and asynchronous networking library, originally developed at FriendFeed.";
       };
     };
-
-
 
     "urllib3" = python.mkDerivation {
       name = "urllib3-1.22";
@@ -580,31 +545,32 @@ let
       };
     };
 
-
-
     "webcolors" = python.mkDerivation {
-      name = "webcolors-1.7";
-      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/1c/11/d9fb5a7c872a941ad8b30a4be191253d5a9028834c4d69eab55bb6bc60be/webcolors-1.7.tar.gz"; sha256 = "e47e68644d41c0b1f1e4d939cfe4039bdf1ab31234df63c7a4f59d4766487206"; };
+      name = "webcolors-1.8.1";
+      src = pkgs.fetchurl { url = "https://pypi.python.org/packages/f5/9c/da29a884572b675b7d3e96aa1a06c6b46138ca58235e834ef94f2660fdd3/webcolors-1.8.1.tar.gz"; sha256 = "030562f624467a9901f0b455fef05486a88cfb5daa1e356bd4aacea043850b59"; };
       doCheck = commonDoCheck;
       buildInputs = commonBuildInputs;
       propagatedBuildInputs = [ ];
       meta = with pkgs.stdenv.lib; {
         homepage = "https://github.com/ubernostrum/webcolors";
         license = licenses.bsdOriginal;
-        description = "A library for working with color names and color value formats defined by the HTML and CSS specifications for use in documents on the Web.";
+        description = "A library for working with color names and color values formats defined by HTML and CSS.";
       };
     };
-
   };
-  overrides = import ./requirements_override.nix { inherit pkgs python; };
+  localOverridesFile = ./requirements_override.nix;
+  overrides = import localOverridesFile { inherit pkgs python; };
   commonOverrides = [
-
+    
   ];
+  allOverrides =
+    (if (builtins.pathExists localOverridesFile)
+     then [overrides] else [] ) ++ commonOverrides;
 
 in python.withPackages
    (fix' (pkgs.lib.fold
             extends
             generated
-            ([overrides] ++ commonOverrides)
+            allOverrides
          )
    )
